Refactor profile/role/quarantine relationship
We add a new subclass of the generated profile class with traits
adding properties and accessors for handling roles and quarantines.

This way we avoid having to stick on public properties runtime.

The profile repository will always return a subclassed profile when
retrieving a single profile and provides an method for enriching
profiles in general as well.

Forms and blocks dealing with a single profile use this subclass.